Move test/utils -> src/utils
