+++ src/benchmarks/utils/formatTime.ts
@@ -1,12 +1,12 @@
 import prettyTime from 'pretty-time'
 
-export function formatNanos(nanoseconds: bigint) {
+export function formatNanos (nanoseconds: bigint) {
   const seconds = nanoseconds / BigInt(Math.pow(10, 9))
   const remainingNs = nanoseconds - seconds * BigInt(Math.pow(10, 9))
   return prettyTime([Number(seconds), Number(remainingNs)], 'ms')
 }
 
-export function formatMs(milliseconds: number) {
+export function formatMs (milliseconds: number) {
   const seconds = milliseconds / Math.pow(10, 3)
   const remainingMs = milliseconds - seconds * Math.pow(10, 3)
   const remainingNs = remainingMs * Math.pow(10, 6)
