+++ test/benchmarks/02-erc20-transfers.test.ts
@@ -3,8 +3,8 @@ import { Wallet } from 'ethers'
 import { Contract } from 'ethers/contract'
 import { parseEther } from 'ethers/utils'
 import { createSimpleChain, SimpleChain, SimpleProvider } from '../../src/chain'
-import { getERC20TransferTransaction } from './utils/transactions'
-import { deployERC20 } from './utils/deploy'
+import { getERC20TransferTransaction } from '../../src/benchmarks/utils/transactions'
+import { deployERC20 } from '../../src/benchmarks/utils/deploy'
 
 describe('ERC20 transfers', () => {
   let chain: SimpleChain
